+++ libs/simdvec/src/main21/java/org/elasticsearch/simdvec/internal/MemorySegmentES92PanamaInt7VectorsScorer.java
@@ -9,26 +9,33 @@
 package org.elasticsearch.simdvec.internal;
 
 import jdk.incubator.vector.ByteVector;
+import jdk.incubator.vector.FloatVector;
 import jdk.incubator.vector.IntVector;
 import jdk.incubator.vector.ShortVector;
 import jdk.incubator.vector.Vector;
+import jdk.incubator.vector.VectorOperators;
 import jdk.incubator.vector.VectorShape;
 import jdk.incubator.vector.VectorSpecies;
 
+import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.store.IndexInput;
+import org.apache.lucene.util.VectorUtil;
 import org.elasticsearch.simdvec.ES92Int7VectorsScorer;
 
 import java.io.IOException;
 import java.lang.foreign.MemorySegment;
+import java.nio.ByteOrder;
 
 import static java.nio.ByteOrder.LITTLE_ENDIAN;
 import static jdk.incubator.vector.VectorOperators.ADD;
 import static jdk.incubator.vector.VectorOperators.B2I;
 import static jdk.incubator.vector.VectorOperators.B2S;
 import static jdk.incubator.vector.VectorOperators.S2I;
+import static org.apache.lucene.index.VectorSimilarityFunction.EUCLIDEAN;
+import static org.apache.lucene.index.VectorSimilarityFunction.MAXIMUM_INNER_PRODUCT;
 
 /** Panamized scorer for 7-bit quantized vectors stored as an {@link IndexInput}. **/
-abstract class MemorySegmentES92FallBackInt7VectorsScorer extends ES92Int7VectorsScorer {
+abstract class MemorySegmentES92PanamaInt7VectorsScorer extends ES92Int7VectorsScorer {
 
     private static final VectorSpecies<Byte> BYTE_SPECIES_64 = ByteVector.SPECIES_64;
     private static final VectorSpecies<Byte> BYTE_SPECIES_128 = ByteVector.SPECIES_128;
@@ -41,8 +48,8 @@ abstract class MemorySegmentES92FallBackInt7VectorsScorer extends ES92Int7Vector
     private static final VectorSpecies<Integer> INT_SPECIES_512 = IntVector.SPECIES_512;
 
     private static final int VECTOR_BITSIZE;
-    protected static final VectorSpecies<Float> FLOAT_SPECIES;
-    protected static final VectorSpecies<Integer> INT_SPECIES;
+    private static final VectorSpecies<Float> FLOAT_SPECIES;
+    private static final VectorSpecies<Integer> INT_SPECIES;
 
     static {
         // default to platform supported bitsize
@@ -53,12 +60,12 @@ abstract class MemorySegmentES92FallBackInt7VectorsScorer extends ES92Int7Vector
 
     protected final MemorySegment memorySegment;
 
-    public MemorySegmentES92FallBackInt7VectorsScorer(IndexInput in, int dimensions, MemorySegment memorySegment) {
+    public MemorySegmentES92PanamaInt7VectorsScorer(IndexInput in, int dimensions, MemorySegment memorySegment) {
         super(in, dimensions);
         this.memorySegment = memorySegment;
     }
 
-    protected long fallbackInt7DotProduct(byte[] q) throws IOException {
+    protected long panamaInt7DotProduct(byte[] q) throws IOException {
         assert dimensions == q.length;
         int i = 0;
         int res = 0;
@@ -147,7 +154,7 @@ abstract class MemorySegmentES92FallBackInt7VectorsScorer extends ES92Int7Vector
         return acc.reduceLanes(ADD);
     }
 
-    protected void fallbackInt7DotProductBulk(byte[] q, int count, float[] scores) throws IOException {
+    protected void panamaInt7DotProductBulk(byte[] q, int count, float[] scores) throws IOException {
         assert dimensions == q.length;
         // only vectorize if we'll at least enter the loop a single time
         if (dimensions >= 16) {
@@ -248,5 +255,73 @@ abstract class MemorySegmentES92FallBackInt7VectorsScorer extends ES92Int7Vector
             }
             scores[iter] = res;
         }
+    }
+
+    protected void applyCorrectionsBulk(
+        float queryLowerInterval,
+        float queryUpperInterval,
+        int queryComponentSum,
+        float queryAdditionalCorrection,
+        VectorSimilarityFunction similarityFunction,
+        float centroidDp,
+        float[] scores
+    ) throws IOException {
+        int limit = FLOAT_SPECIES.loopBound(BULK_SIZE);
+        int i = 0;
+        long offset = in.getFilePointer();
+        float ay = queryLowerInterval;
+        float ly = (queryUpperInterval - ay) * SEVEN_BIT_SCALE;
+        float y1 = queryComponentSum;
+        for (; i < limit; i += FLOAT_SPECIES.length()) {
+            var ax = FloatVector.fromMemorySegment(FLOAT_SPECIES, memorySegment, offset + i * Float.BYTES, ByteOrder.LITTLE_ENDIAN);
+            var lx = FloatVector.fromMemorySegment(
+                FLOAT_SPECIES,
+                memorySegment,
+                offset + 4 * BULK_SIZE + i * Float.BYTES,
+                ByteOrder.LITTLE_ENDIAN
+            ).sub(ax).mul(SEVEN_BIT_SCALE);
+            var targetComponentSums = IntVector.fromMemorySegment(
+                INT_SPECIES,
+                memorySegment,
+                offset + 8 * BULK_SIZE + i * Integer.BYTES,
+                ByteOrder.LITTLE_ENDIAN
+            ).convert(VectorOperators.I2F, 0);
+            var additionalCorrections = FloatVector.fromMemorySegment(
+                FLOAT_SPECIES,
+                memorySegment,
+                offset + 12 * BULK_SIZE + i * Float.BYTES,
+                ByteOrder.LITTLE_ENDIAN
+            );
+            var qcDist = FloatVector.fromArray(FLOAT_SPECIES, scores, i);
+            // ax * ay * dimensions + ay * lx * (float) targetComponentSum + ax * ly * y1 + lx * ly *
+            // qcDist;
+            var res1 = ax.mul(ay).mul(dimensions);
+            var res2 = lx.mul(ay).mul(targetComponentSums);
+            var res3 = ax.mul(ly).mul(y1);
+            var res4 = lx.mul(ly).mul(qcDist);
+            var res = res1.add(res2).add(res3).add(res4);
+            // For euclidean, we need to invert the score and apply the additional correction, which is
+            // assumed to be the squared l2norm of the centroid centered vectors.
+            if (similarityFunction == EUCLIDEAN) {
+                res = res.mul(-2).add(additionalCorrections).add(queryAdditionalCorrection).add(1f);
+                res = FloatVector.broadcast(FLOAT_SPECIES, 1).div(res).max(0);
+                res.intoArray(scores, i);
+            } else {
+                // For cosine and max inner product, we need to apply the additional correction, which is
+                // assumed to be the non-centered dot-product between the vector and the centroid
+                res = res.add(queryAdditionalCorrection).add(additionalCorrections).sub(centroidDp);
+                if (similarityFunction == MAXIMUM_INNER_PRODUCT) {
+                    res.intoArray(scores, i);
+                    // not sure how to do it better
+                    for (int j = 0; j < FLOAT_SPECIES.length(); j++) {
+                        scores[i + j] = VectorUtil.scaleMaxInnerProductScore(scores[i + j]);
+                    }
+                } else {
+                    res = res.add(1f).mul(0.5f).max(0);
+                    res.intoArray(scores, i);
+                }
+            }
+        }
+        in.seek(offset + 16L * BULK_SIZE);
     }
 }
